+++ siteapp/urls.py
@@ -17,6 +17,7 @@ admin.autodiscover()
 import siteapp.views as views
 import siteapp.views_landing as views_landing
 import siteapp.views_health as views_health
+import siteapp.views_management as views_management
 from .good_settings_helpers import signup_wrapper
 from .settings import *
 
@@ -151,6 +152,11 @@ urlpatterns = [
 
     # Session
     url(r'session_security/', include('session_security.urls')),
+
+    # management commands
+    url(r'^management/$', views_management.index),
+    url(r'^management/is_superuser$', views_management.is_superuser),
+    url(r'^management/listcomponents$', views_management.listcomponents),
 ]
 
 urlpatterns += [url(r'^api/v2/', include('api.urls'))]