+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 GovReady-Q Release Notes
 ========================
 
+v0.11.6 (March 14, 2023)
+---------------------------
+
+**Developer changes**
+
+* Upgrade Python libraries.
+
+**Developer changes**
+
+* Add siteapp.management_views as webhooks for calling Django management commands.
+
+
 v0.11.5 (March 2, 2023)
 ---------------------------
 
@@ -24,7 +36,6 @@ v0.11.3 (December 10, 2022)
 **Developer changes**
 
 * Add processing for question actions targeted at system to handle `system/add_baseline/<value>` to add additional baseline set of controls to a system without deleting already assigned controls.A
-
 
 
 v0.11.2 (December 10, 2022)